+++ xrGame/ui/UIListWnd_inline.h
@@ -38,7 +38,7 @@ bool CUIListWnd::AddItem(Element *pItem, int insertBeforeIdx)
 	if (-1 == insertBeforeIdx)
 	{
 		m_ItemList.push_back(pItem);
-		pItem->SetIndex(m_ItemList.size() - 1);
+		pItem->SetIndex(static_cast<int>(m_ItemList.size() - 1));
 	}
 	else
 	{
